[#114] refactor: Swagger Tag 변경과 Operation 추가
